feat: implement fastify handler class and set error handler method
