+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@secjs/http",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "",
   "license": "MIT",
   "author": "João Lenon",
